Added Tags store.
